add AnimationControl
